+++ src/PKSim.Presentation/DTO/Mappers/ParameterToParameterDTOMapper.cs
@@ -9,6 +9,7 @@ using OSPSuite.Presentation.DTO;
 using PKSim.Core;
 using PKSim.Core.Mappers;
 using PKSim.Core.Repositories;
+using PKSim.Presentation.DTO.Compounds;
 using PKSim.Presentation.DTO.Parameters;
 
 namespace PKSim.Presentation.DTO.Mappers
@@ -16,6 +17,7 @@ namespace PKSim.Presentation.DTO.Mappers
    public interface IParameterToParameterDTOMapper : OSPSuite.Presentation.Mappers.IParameterToParameterDTOMapper
    {
       IParameterDTO MapAsReadWriteFrom(IParameter parameter);
+      MolWeightParameterDTO MapMolWeightDTOFrom(IParameter molWeight, IParameter effectiveMolWeight);
    }
 
    public class ParameterToParameterDTOMapper : IParameterToParameterDTOMapper
@@ -50,7 +52,7 @@ namespace PKSim.Presentation.DTO.Mappers
             return new NullParameterDTO();
 
          var parameterDTO = new ParameterDTO(parameter);
-         UpdateParameterDTOFromParameter(parameterDTO, parameter);
+         updateParameterDTOFromParameter(parameterDTO, parameter);
          return parameterDTO;
       }
 
@@ -60,11 +62,18 @@ namespace PKSim.Presentation.DTO.Mappers
             return new NullParameterDTO();
 
          var parameterDTO = new WritableParameterDTO(parameter);
-         UpdateParameterDTOFromParameter(parameterDTO, parameter);
+         updateParameterDTOFromParameter(parameterDTO, parameter);
          return parameterDTO;
       }
 
-      protected void UpdateParameterDTOFromParameter(ParameterDTO parameterDTO, IParameter parameter)
+      public MolWeightParameterDTO MapMolWeightDTOFrom(IParameter molWeight, IParameter effectiveMolWeight)
+      {
+         var molWeightParameterDTO = new MolWeightParameterDTO(molWeight, effectiveMolWeight);
+         updateParameterDTOFromParameter(molWeightParameterDTO, molWeight);
+         return molWeightParameterDTO;
+      }
+
+      private void updateParameterDTOFromParameter(ParameterDTO parameterDTO, IParameter parameter)
       {
          var parameterPath = _entityPathResolver.ObjectPathFor(parameter);
          var representationInfo = _representationInfoRepository.InfoFor(parameter);
